wip: move string buffer to begin collection
to simplify the way GetComments and GetAllComments were (not)working, instead there's a BeginCollection which passes in the string buffer for the new collection's comment block. the collection can read from that whenever it wants ( ie. on finalize value )
all notes and decode tests pass; now on to the file/comments tests