first pass at handling yaml blocks
the triple quotes heredocs break yaml validation.
unless tell gets its own validators, its heredocs are... aspirational.
until then: support a leading pipe (|) as an "ambiguous" heredoc -- that depends on the closing quotes to know whether to escape and how to indent the block.
it's not yaml -- but its compatible enough with it for validation.