charm cursor: don't exit on eof
charm cursor is meant to be more of a "watcher" it shouldn't choose any particular action on Eof, instead it can just return "unhandled". when in a parallel state, this will cause it to be dropped ( where previously it would have forced a terminal transition ) -- but at Eof the machine is exiting, so that's fine.